+++ lib/plane.rb
